Add nfr24l01 protocol

+++ Firmware/Arduino/src/intr.h
@@ -1,7 +1,7 @@
 #ifndef INTR_H
 #define INTR_H
 
-extern volatile int intr_fired;
+extern volatile uint8_t intr_fired;
 
 extern void intr_init();
